LPS-96987 Move the upgrade to the Core since the modified tables are defined there
