+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/upgrade/LayoutImplUpgrade.java
@@ -15,7 +15,7 @@
 package com.liferay.layout.internal.upgrade;
 
 import com.liferay.layout.internal.upgrade.v1_0_0.UpgradeLayoutPermissions;
-import com.liferay.layout.internal.upgrade.v1_0_1.UpgradeLayoutParentPlid;
+import com.liferay.portal.upgrade.v7_2_x.UpgradeLayoutParentPlid;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 
 import org.osgi.service.component.annotations.Component;
@@ -29,7 +29,6 @@ public class LayoutImplUpgrade implements UpgradeStepRegistrator {
 	@Override
 	public void register(Registry registry) {
 		registry.register("0.0.0", "1.0.0", new UpgradeLayoutPermissions());
-		registry.register("1.0.0", "1.0.1", new UpgradeLayoutParentPlid());
 	}
 
 }
