+++ portal-impl/src/com/liferay/portal/upgrade/v7_2_x/PortalUpgradeProcessRegistryImpl.java
@@ -40,6 +40,8 @@ public class PortalUpgradeProcessRegistryImpl
 		upgradeProcesses.put(new Version(5, 0, 1), new UpgradePersonalMenu());
 
 		upgradeProcesses.put(new Version(5, 0, 2), new UpgradeCountry());
+
+		upgradeProcesses.put(new Version(5, 0, 3), new UpgradeLayoutParentPlid());
 	}
 
 }
